import styles from styles.ts
